extract types to shard and update fees pane padding
