+++ extension/src/popup/components/InternalTransaction/FeesPane/styles.scss
@@ -4,7 +4,7 @@
   display: flex;
   flex-direction: column;
   width: 100%;
-  padding: pxToRem(32px) 0;
+  padding: 0 0;
   gap: pxToRem(16px);
 
   &__Header {
